removed urls to constants file.
